Modules: move connection_control to a module, move transport to a module
